+++ src/Stories/FloatingButton.story.lua
@@ -1,16 +1,14 @@
 local Hoarcekat = script:FindFirstAncestor("Hoarcekat")
-
-local Assets = require(Hoarcekat.Plugin.Assets)
-local FloatingButton = require(script.Parent.FloatingButton)
 local Roact = require(Hoarcekat.Vendor.Roact)
-
-local e = Roact.createElement
+local Assets = require(Hoarcekat.Plugin.Assets)
+local FloatingButton = require(Hoarcekat.Plugin.Components.FloatingButton)
 
 local function TestFloatingButton()
-	return e(FloatingButton, {
+	return Roact.createElement(FloatingButton, {
 		Activated = function()
 			print("activated!")
 		end,
+
 		Image = Assets.preview,
 		ImageSize = UDim.new(0, 24),
 		Size = UDim.new(0, 40),
@@ -18,7 +16,7 @@ local function TestFloatingButton()
 end
 
 return function(target)
-	local handle = Roact.mount(e(TestFloatingButton), target, "FloatingButton")
+	local handle = Roact.mount(Roact.createElement(TestFloatingButton, {}), target, "FloatingButton")
 
 	return function()
 		Roact.unmount(handle)