+++ CHANGELOG.md
@@ -4,6 +4,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.2.3]
+### Added
+- Added a way to toggle pattern searching in the search bar.
+- Added my missing `stylua.toml` file.
+
+### Changed
+- Moved stories to a story folder.
+
+### Fixed
+- Fixed an error on plugin update.
+
 ## [1.2.2]
 ### Added
 - Added a hook library.